+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Magic Formula Strategy
+- Mean Reversion Strategy
+- Allow different window sizes for same return engine in return_engine_dict
+- Allow different window sized for return and risk engine
 ### Fixed
 ### Changed
 ### Removed